+++ src/optimizers/AdamOptimizer.cpp
@@ -28,11 +28,11 @@ namespace kann
     const Tensor& v = variable.optimizer_states[1];
 
     assert(variable.gradient);
-    Tensor m_new = math::cwise(m, *variable.gradient, [this](double m, double gradient) { return m_beta1 * m + (1.0-m_beta1) * gradient * gradient; });
-    Tensor v_new = math::cwise(v, *variable.gradient, [this](double v, double gradient) { return m_beta2 * v + (1.0-m_beta2) * gradient; });
+    Tensor m_new = math::cwise(m, *variable.gradient, [this](float m, float gradient) { return m_beta1 * m + (1.0-m_beta1) * gradient * gradient; });
+    Tensor v_new = math::cwise(v, *variable.gradient, [this](float v, float gradient) { return m_beta2 * v + (1.0-m_beta2) * gradient; });
 
-    Tensor m_hat = math::scale(m_new, 1.0 / (1.0 - std::pow(m_beta1, (double)m_timestep)));
-    Tensor v_hat = math::scale(m_new, 1.0 / (1.0 - std::pow(m_beta2, (double)m_timestep)));
+    Tensor m_hat = math::scale(m_new, 1.0 / (1.0 - std::pow(m_beta1, (float)m_timestep)));
+    Tensor v_hat = math::scale(m_new, 1.0 / (1.0 - std::pow(m_beta2, (float)m_timestep)));
 
     Tensor correction = math::scale(m_hat, m_alpha / ( math::norm(v_hat.as_ref()) + m_epsilon ));
     variable.value = math::sub(variable.value, correction);
